requirements.txt
asgiref==3.3.1
Django==3.1.5
django-phonenumber-field==5.0.0
form==0.0.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
